+++ backend/src/main/java/eu/zuinnote/example/springwebdemo/configuration/application/oidc/OidcMapper.java
@@ -1,12 +1,16 @@
 package eu.zuinnote.example.springwebdemo.configuration.application.oidc;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import org.springframework.validation.annotation.Validated;
 
 @Validated
 public class OidcMapper {
     private String authoritiesPrefix;
-    private ArrayList<String> jwtRoleClaims;
+    private ArrayList<String> jwtIdTokenClaims;
+    private ArrayList<String> userClaims;
+    private ArrayList<String> userAttributes;
+    private HashMap<String, String> claimsSeparatorMap;
 
     public String getAuthoritiesPrefix() {
         return this.authoritiesPrefix;
@@ -16,11 +20,35 @@ public class OidcMapper {
         this.authoritiesPrefix = authoritiesPrefix;
     }
 
-    public ArrayList<String> getJwtRoleClaims() {
-        return this.jwtRoleClaims;
+    public HashMap<String, String> getClaimsSeparatorMap() {
+        return this.claimsSeparatorMap;
     }
 
-    public void setJwtRoleClaims(ArrayList<String> jwtRoleClaims) {
-        this.jwtRoleClaims = jwtRoleClaims;
+    public void setClaimsSeparatorMap(HashMap<String, String> claimsSeparatorMap) {
+        this.claimsSeparatorMap = claimsSeparatorMap;
+    }
+
+    public ArrayList<String> getJwtIdTokenClaims() {
+        return this.jwtIdTokenClaims;
+    }
+
+    public void setJwtIdTokenClaims(ArrayList<String> jwtIdTokenClaims) {
+        this.jwtIdTokenClaims = jwtIdTokenClaims;
+    }
+
+    public ArrayList<String> getUserClaims() {
+        return this.userClaims;
+    }
+
+    public void setUserClaims(ArrayList<String> userClaims) {
+        this.userClaims = userClaims;
+    }
+
+    public ArrayList<String> getUserAttributes() {
+        return this.userAttributes;
+    }
+
+    public void setUserAttributes(ArrayList<String> userAttributes) {
+        this.userAttributes = userAttributes;
     }
 }